@@ -3,7 +3,11 @@ var StringBinding = require('sharedb-string-binding');
 
 // Open WebSocket connection to ShareDB server
 var ReconnectingWebSocket = require('reconnecting-websocket');
-var socket = new ReconnectingWebSocket('ws://' + window.location.host);
+var socket = new ReconnectingWebSocket('ws://' + window.location.host, [], {
+  // ShareDB handles dropped messages, and buffering them while the socket
+  // is closed has undefined behavior
+  maxEnqueuedMessages: 0
+});
 var connection = new sharedb.Connection(socket);
 
 var element = document.querySelector('textarea');
