@@ -1,8 +1,9 @@
 var hat = require('hat');
+var ShareDBError = require('./error');
+var logger = require('./logger');
+var ACTIONS = require('./message-actions').ACTIONS;
 var types = require('./types');
 var util = require('./util');
-var logger = require('./logger');
-var ShareDBError = require('./error');
 
 var ERROR_CODE = ShareDBError.CODES;
 
@@ -57,8 +58,13 @@ function Agent(backend, stream) {
   // active, and it is passed to each middleware call
   this.custom = {};
 
+  // The first message received over the connection. Stored to warn if messages
+  // are being sent before the handshake.
+  this._firstReceivedMessage = null;
+  this._handshakeReceived = false;
+
   // Send the legacy message to initialize old clients with the random agent Id
-  this.send(this._initMessage('init'));
+  this.send(this._initMessage(ACTIONS.initLegacy));
 }
 module.exports = Agent;
 
@@ -174,7 +180,7 @@ Agent.prototype._subscribeToQuery = function(emitter, queryId, collection, query
 
   var agent = this;
   emitter.onExtra = function(extra) {
-    agent.send({a: 'q', id: queryId, extra: extra});
+    agent.send({a: ACTIONS.queryUpdate, id: queryId, extra: extra});
   };
 
   emitter.onDiff = function(diff) {
@@ -186,7 +192,7 @@ Agent.prototype._subscribeToQuery = function(emitter, queryId, collection, query
     }
     // Consider stripping the collection out of the data we send here
     // if it matches the query's collection.
-    agent.send({a: 'q', id: queryId, diff: diff});
+    agent.send({a: ACTIONS.queryUpdate, id: queryId, diff: diff});
   };
 
   emitter.onError = function(err) {
@@ -250,7 +256,7 @@ Agent.prototype.send = function(message) {
 
 Agent.prototype._sendOp = function(collection, id, op) {
   var message = {
-    a: 'op',
+    a: ACTIONS.op,
     c: collection,
     d: id,
     v: op.v,
@@ -354,22 +360,34 @@ Agent.prototype._open = function() {
 
 // Check a request to see if its valid. Returns an error if there's a problem.
 Agent.prototype._checkRequest = function(request) {
-  if (request.a === 'qf' || request.a === 'qs' || request.a === 'qu') {
+  if (
+    request.a === ACTIONS.queryFetch ||
+    request.a === ACTIONS.querySubscribe ||
+    request.a === ACTIONS.queryUnsubscribe
+  ) {
     // Query messages need an ID property.
     if (typeof request.id !== 'number') return 'Missing query ID';
-  } else if (request.a === 'op' || request.a === 'f' || request.a === 's' || request.a === 'u' || request.a === 'p') {
+  } else if (request.a === ACTIONS.op ||
+      request.a === ACTIONS.fetch ||
+      request.a === ACTIONS.subscribe ||
+      request.a === ACTIONS.unsubscribe ||
+      request.a === ACTIONS.presence) {
     // Doc-based request.
     if (request.c != null && typeof request.c !== 'string') return 'Invalid collection';
     if (request.d != null && typeof request.d !== 'string') return 'Invalid id';
 
-    if (request.a === 'op' || request.a === 'p') {
+    if (request.a === ACTIONS.op || request.a === ACTIONS.presence) {
       if (request.v != null && (typeof request.v !== 'number' || request.v < 0)) return 'Invalid version';
     }
 
-    if (request.a === 'p') {
+    if (request.a === ACTIONS.presence) {
       if (typeof request.id !== 'string') return 'Missing presence ID';
     }
-  } else if (request.a === 'bf' || request.a === 'bs' || request.a === 'bu') {
+  } else if (
+    request.a === ACTIONS.bulkFetch ||
+    request.a === ACTIONS.bulkSubscribe ||
+    request.a === ACTIONS.bulkUnsubscribe
+  ) {
     // Bulk request
     if (request.c != null && typeof request.c !== 'string') return 'Invalid collection';
     if (typeof request.b !== 'object') return 'Invalid bulk subscribe data';
@@ -381,30 +399,31 @@ Agent.prototype._handleMessage = function(request, callback) {
   try {
     var errMessage = this._checkRequest(request);
     if (errMessage) return callback(new ShareDBError(ERROR_CODE.ERR_MESSAGE_BADLY_FORMED, errMessage));
+    this._checkFirstMessage(request);
 
     switch (request.a) {
-      case 'hs':
+      case ACTIONS.handshake:
         if (request.id) this.src = request.id;
-        return callback(null, this._initMessage('hs'));
-      case 'qf':
+        return callback(null, this._initMessage(ACTIONS.handshake));
+      case ACTIONS.queryFetch:
         return this._queryFetch(request.id, request.c, request.q, getQueryOptions(request), callback);
-      case 'qs':
+      case ACTIONS.querySubscribe:
         return this._querySubscribe(request.id, request.c, request.q, getQueryOptions(request), callback);
-      case 'qu':
+      case ACTIONS.queryUnsubscribe:
         return this._queryUnsubscribe(request.id, callback);
-      case 'bf':
+      case ACTIONS.bulkFetch:
         return this._fetchBulk(request.c, request.b, callback);
-      case 'bs':
+      case ACTIONS.bulkSubscribe:
         return this._subscribeBulk(request.c, request.b, callback);
-      case 'bu':
+      case ACTIONS.bulkUnsubscribe:
         return this._unsubscribeBulk(request.c, request.b, callback);
-      case 'f':
+      case ACTIONS.fetch:
         return this._fetch(request.c, request.d, request.v, callback);
-      case 's':
+      case ACTIONS.subscribe:
         return this._subscribe(request.c, request.d, request.v, callback);
-      case 'u':
+      case ACTIONS.unsubscribe:
         return this._unsubscribe(request.c, request.d, callback);
-      case 'op':
+      case ACTIONS.op:
         // Normalize the properties submitted
         var op = createClientOp(request, this._src());
         if (op.seq >= util.MAX_SAFE_INTEGER) {
@@ -415,11 +434,11 @@ Agent.prototype._handleMessage = function(request, callback) {
         }
         if (!op) return callback(new ShareDBError(ERROR_CODE.ERR_MESSAGE_BADLY_FORMED, 'Invalid op message'));
         return this._submit(request.c, request.d, op, callback);
-      case 'nf':
+      case ACTIONS.snapshotFetch:
         return this._fetchSnapshot(request.c, request.d, request.v, callback);
-      case 'nt':
+      case ACTIONS.snapshotFetchByTimestamp:
         return this._fetchSnapshotByTimestamp(request.c, request.d, request.ts, callback);
-      case 'p':
+      case ACTIONS.presence:
         if (!this.backend.presenceEnabled) return;
         var presence = this._createPresence(request);
         if (presence.t && !util.supportsPresence(types.map[presence.t])) {
@@ -429,11 +448,13 @@ Agent.prototype._handleMessage = function(request, callback) {
           });
         }
         return this._broadcastPresence(presence, callback);
-      case 'ps':
+      case ACTIONS.presenceSubscribe:
         if (!this.backend.presenceEnabled) return;
         return this._subscribePresence(request.ch, request.seq, callback);
-      case 'pu':
+      case ACTIONS.presenceUnsubscribe:
         return this._unsubscribePresence(request.ch, request.seq, callback);
+      case ACTIONS.pingPong:
+        return this._pingPong(callback);
       default:
         callback(new ShareDBError(ERROR_CODE.ERR_MESSAGE_BADLY_FORMED, 'Invalid or unknown message'));
     }
@@ -525,6 +546,14 @@ Agent.prototype._querySubscribe = function(queryId, collection, query, options, 
     };
     finish();
   });
+};
+
+Agent.prototype._pingPong = function(callback) {
+  var error = null;
+  var message = {
+    a: ACTIONS.pingPong
+  };
+  callback(error, message);
 };
 
 function getResultsData(results) {
@@ -695,9 +724,10 @@ Agent.prototype._unsubscribeBulk = function(collection, ids, callback) {
 
 Agent.prototype._submit = function(collection, id, op, callback) {
   var agent = this;
-  this.backend.submit(this, collection, id, op, null, function(err, ops) {
+  this.backend.submit(this, collection, id, op, null, function(err, ops, request) {
     // Message to acknowledge the op was successfully submitted
     var ack = {src: op.src, seq: op.seq, v: op.v};
+    if (request._fixupOps.length) ack[ACTIONS.fixup] = request._fixupOps;
     if (err) {
       // Occasional 'Op already submitted' errors are expected to happen as
       // part of normal operation, since inflight ops need to be resent after
@@ -761,7 +791,7 @@ Agent.prototype._broadcastPresence = function(presence, callback) {
 
 Agent.prototype._createPresence = function(request) {
   return {
-    a: 'p',
+    a: ACTIONS.presence,
     ch: request.ch,
     src: this._src(),
     id: request.id, // Presence ID, not Doc ID (which is 'd')
@@ -813,7 +843,7 @@ Agent.prototype._requestPresence = function(channel, callback) {
 Agent.prototype._handlePresenceData = function(presence) {
   if (presence.src === this._src()) return;
 
-  if (presence.r) return this.send({a: 'pr', ch: presence.ch});
+  if (presence.r) return this.send({a: ACTIONS.presenceRequest, ch: presence.ch});
 
   var backend = this.backend;
   var context = {
@@ -824,11 +854,25 @@ Agent.prototype._handlePresenceData = function(presence) {
   backend.trigger(backend.MIDDLEWARE_ACTIONS.sendPresence, this, context, function(error) {
     if (error) {
       if (backend.doNotForwardSendPresenceErrorsToClient) backend.errorHandler(error, {agent: agent});
-      else agent.send({a: 'p', ch: presence.ch, id: presence.id, error: getReplyErrorObject(error)});
+      else agent.send({a: ACTIONS.presence, ch: presence.ch, id: presence.id, error: getReplyErrorObject(error)});
       return;
     }
     agent.send(presence);
   });
+};
+
+Agent.prototype._checkFirstMessage = function(request) {
+  if (this._handshakeReceived) return;
+  if (!this._firstReceivedMessage) this._firstReceivedMessage = request;
+
+  if (request.a === ACTIONS.handshake) {
+    this._handshakeReceived = true;
+    if (this._firstReceivedMessage.a !== ACTIONS.handshake) {
+      logger.warn('Unexpected message received before handshake', this._firstReceivedMessage);
+    }
+    // Release memory
+    this._firstReceivedMessage = null;
+  }
 };
 
 function createClientOp(request, clientId) {
